refactor: server improvement
